Create an Example on how this blog can be used with boomerang
It is still quite inconvenient as far as setting it up goes. But that
may well change in the (near) future.

However, the implementation makes the dependency of saeplog to happstack almost
optional. For the rapid development stage that this project is still in, it
does not make a lot of sense to split the source into seperate packages, just
to keep the dependencies small. To actually run and test everything, it is
necessary to hava all those dependencies installed anyway.